Code: shrink GmsSupport. Code: fake-signature support.
