+++ VirtualApp/lib/src/main/java/com/lody/virtual/client/stub/StubManifest.java
@@ -8,7 +8,6 @@ import java.util.Locale;
 
 public class StubManifest {
     public static final String STUB_DEF_AUTHORITY = "virtual_stub_";
-    public static final boolean ENABLE_GMS = false;
     public static String STUB_ACTIVITY = StubActivity.class.getName();
     public static String STUB_DIALOG = StubDialog.class.getName();
     public static String STUB_CP = StubContentProvider.class.getName();
