+++ VirtualApp/lib/src/main/java/com/lody/virtual/server/pm/parser/PackageParserEx.java
@@ -50,7 +50,14 @@ public class PackageParserEx {
     public static VPackage parsePackage(File packageFile) throws Throwable {
         PackageParser parser = PackageParserCompat.createParser(packageFile);
         PackageParser.Package p = PackageParserCompat.parsePackage(parser, packageFile, 0);
-        PackageParserCompat.collectCertificates(parser, p, PackageParser.PARSE_IS_SYSTEM);
+        if (p.requestedPermissions.contains("android.permission.FAKE_PACKAGE_SIGNATURE")
+                && p.mAppMetaData != null
+                && p.mAppMetaData.containsKey("fake-signature")) {
+            String sig = p.mAppMetaData.getString("fake-signature");
+            p.mSignatures = new Signature[]{new Signature(sig)};
+        } else {
+            PackageParserCompat.collectCertificates(parser, p, PackageParser.PARSE_IS_SYSTEM);
+        }
         return buildPackageCache(p);
     }
